+++ src/icalendar/caselessdict.py
@@ -2,8 +2,12 @@ from icalendar.parser_tools import to_unicode
 
 from collections import OrderedDict
 
+from typing import Any, Optional, Iterable, Mapping, TypeVar
 
-def canonsort_keys(keys, canonical_order=None):
+KT = TypeVar("KT")
+VT = TypeVar("VT")
+
+def canonsort_keys(keys: Iterable[KT], canonical_order: Optional[Iterable[KT]] = None) -> list[KT]:
     """Sorts leading keys according to canonical_order.  Keys not specified in
     canonical_order will appear alphabetically at the end.
     """
@@ -13,7 +17,7 @@ def canonsort_keys(keys, canonical_order=None):
     return sorted(head, key=lambda k: canonical_map[k]) + sorted(tail)
 
 
-def canonsort_items(dict1, canonical_order=None):
+def canonsort_items(dict1: Mapping[KT, VT], canonical_order: Optional[Iterable[KT]] = None) -> list[tuple[KT, VT]]:
     """Returns a list of items from dict1, sorted by canonical_order."""
     return [(k, dict1[k]) for k in canonsort_keys(dict1.keys(), canonical_order)]
 
@@ -23,7 +27,7 @@ class CaselessDict(OrderedDict):
     Values retain their case.
     """
 
-    def __init__(self, *args, **kwargs):
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
         """Set keys to upper for initial dict."""
         super().__init__(*args, **kwargs)
         for key, value in self.items():
@@ -32,42 +36,42 @@ class CaselessDict(OrderedDict):
                 super().__delitem__(key)
                 self[key_upper] = value
 
-    def __getitem__(self, key):
+    def __getitem__(self, key: Any) -> Any:
         key = to_unicode(key)
         return super().__getitem__(key.upper())
 
-    def __setitem__(self, key, value):
+    def __setitem__(self, key: Any, value: Any) -> None:
         key = to_unicode(key)
         super().__setitem__(key.upper(), value)
 
-    def __delitem__(self, key):
+    def __delitem__(self, key: Any) -> None:
         key = to_unicode(key)
         super().__delitem__(key.upper())
 
-    def __contains__(self, key):
+    def __contains__(self, key: Any) -> bool:
         key = to_unicode(key)
         return super().__contains__(key.upper())
 
-    def get(self, key, default=None):
+    def get(self, key: Any, default: Optional[Any] = None) -> Any:
         key = to_unicode(key)
         return super().get(key.upper(), default)
 
-    def setdefault(self, key, value=None):
+    def setdefault(self, key: Any, value: Optional[Any] = None) -> Any:
         key = to_unicode(key)
         return super().setdefault(key.upper(), value)
 
-    def pop(self, key, default=None):
+    def pop(self, key: Any, default: Optional[Any] = None) -> Any:
         key = to_unicode(key)
         return super().pop(key.upper(), default)
 
-    def popitem(self):
+    def popitem(self) -> tuple[str, Any]:
         return super().popitem()
 
-    def has_key(self, key):
+    def has_key(self, key: Any) -> bool:
         key = to_unicode(key)
         return super().__contains__(key.upper())
 
-    def update(self, *args, **kwargs):
+    def update(self, *args: Any, **kwargs: Any) -> None:
         # Multiple keys where key1.upper() == key2.upper() will be lost.
         mappings = list(args) + [kwargs]
         for mapping in mappings:
@@ -76,29 +80,29 @@ class CaselessDict(OrderedDict):
             for key, value in mapping:
                 self[key] = value
 
-    def copy(self):
+    def copy(self) -> "CaselessDict":
         return type(self)(super().copy())
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         return f"{type(self).__name__}({dict(self)})"
 
-    def __eq__(self, other):
+    def __eq__(self, other: object) -> bool:
         return self is other or dict(self.items()) == dict(other.items())
 
-    def __ne__(self, other):
+    def __ne__(self, other: object) -> bool:
         return not self == other
 
     # A list of keys that must appear first in sorted_keys and sorted_items;
     # must be uppercase.
     canonical_order = None
 
-    def sorted_keys(self):
+    def sorted_keys(self) -> list[str]:
         """Sorts keys according to the canonical_order for the derived class.
         Keys not specified in canonical_order will appear at the end.
         """
         return canonsort_keys(self.keys(), self.canonical_order)
 
-    def sorted_items(self):
+    def sorted_items(self) -> list[tuple[str, Any]]:
         """Sorts items according to the canonical_order for the derived class.
         Items not specified in canonical_order will appear at the end.
         """