+++ CHANGES.rst
@@ -26,6 +26,7 @@ Minor changes:
 
 - Move sections in Table of Content of Reference guide.
 - Improve :py:class:`icalendar.prop.vDatetime` documentation. See `Issue #946 <https://github.com/collective/icalendar/issues/946>`_.
+- Add type hints to ``icalendar.caselessdict`` file. See `Issue #938 <https://github.com/collective/icalendar/issues/938>`_.
 
 Breaking changes: